+++ starsim/__init__.py
@@ -1,7 +1,7 @@
 from .version import __version__, __versiondate__, __license__
 from .settings      import *
 from .utils         import *
-from .time          import *
+from .timeunits     import *
 from .parameters    import *
 from .distributions import *
 from .states        import *
