+++ tests/update_baseline.py
@@ -1,8 +1,7 @@
 #!/usr/bin/env python3
-
-'''
+"""
 Run this script to regenerate the baseline.
-'''
+"""
 
 import test_baselines as tb
